load threads in My Threads menu
